pyproject.toml
[tool.poetry]
name = "tomoro"
version = "0.1.0"
description = "RAG system with evaluation"
authors = ["Mehdi HM <m.h.moghadam1996@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
openai = "^1.47.0"
langchain = "^0.2.10"  # Downgraded langchain to 0.2.x to maintain compatibility
rank-bm25 = "^0.2.2"
scikit-learn = "^1.2.2"
nltk = "^3.8.1"
ragas = "^0.1.20"
datasets = "^3.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.7.2"
tqdm = "^4.66.1"
httpx = "^0.24.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"